requirements.txt
gTTS
PyTelegramBotAPI
ffmpeg-python
